--- src/templates/csr/fabric-custom/test.json
@@ -1,7 +1,0 @@
-{
-  "DapAssetsFolder": "GooglePixel2",
-  "ThirdPartyTag": "https://tpc.googlesyndication.com/pagead/imgad?id=CICAgKCrtKGjoAEQARgBMgjYagW3ARF6Fw",
-  "TrackingPixel": "https://gu.com/tracking-pixel.gif",
-  "ResearchPixel": "https://gu.com/research-pixel.gif",
-  "ViewabilityTracker": "<img src='https://gu.com/viewability-tracker.gif'>"
-}
